python/salesforce/requirements.txt
pymssql
python-dotenv
salesforce_api
simple-salesforce
slack_sdk
